Proper latency measuring
Start measuring only when queue task is started
